Add protocol description to connection manager file
Summary:
KDE-Telepathy was complaining that connectcm didn't implement the kdeconnect protocol. Well, now it says it does. This allows KTP to automatically integrate with the connectcm plugin and display incoming SMS.

There seems to be a weirdness on launching the plugin for the very first time, in that one needs to manually launch it because otherwise Telepathy has no kdeconnect account, so it doesn't try to launch the plugin

Test Plan:
I am running Fedora 25 with Plasma -- People on other distros and desktop environments should give this patch a test.

Procedure:
Download this repository, run cmake (make sure to set the install prefix appropriately), run make and make install, then reboot, make sure KDE Connect is connected, and send yourself an SMS.
If the fallback of notification + reply button appears, try manually running connectcm one time to make sure an account is registered.

Expected Results:
Telepathy should handle the request and a notification or message box should appear

In case of failure:
Run ktp-debugger and look for a tab called Kdeconnect. If that is there, it indicates the plugin has loaded but something else has gone wrong.
Regardless of the Kdeconnect tab, run ktp-debugger, open the Mission Control tab, inspect the log for sensitive data, then save it.

Reviewers: #kde_connect, davidedmundson

Reviewed By: davidedmundson

Subscribers: davidedmundson, sredman

Tags: #kde_connect

Differential Revision: https://phabricator.kde.org/D5827